app-producer.py
import asyncio
import websockets
from quixstreams import Application
import os
import json

KAFKA_BROKER = os.getenv("KAFKA_BROKER", "127.0.0.1:19092,127.0.0.1:29092,127.0.0.1:39092")
uri = "wss://jetstream2.us-west.bsky.network/subscribe?wantedCollections=app.bsky.feed.post"

async def listen_to_bluesky():

    app = Application(
        broker_address=KAFKA_BROKER,
        # consumer_group="wikipedia-producer",
        producer_extra_config={
            # Resolve localhost to 127.0.0.1 to avoid IPv6 issues
            "broker.address.family": "v4",
        }
    )   
    topic = app.topic(
        name="bluesky-events",
        value_serializer="json",
    )

    while True:
      try:
        async with websockets.connect(
            uri,
            ping_interval=20,  # Send ping every 20 seconds
            ping_timeout=60,   # Wait 60 seconds for pong response
            close_timeout=10
        ) as websocket:
          print("Connected to Bluesky firehose...")
          while True:
            try:
              message = await websocket.recv()
              data = json.loads(message)
              print(data)

              # Use the event ID as the key for partitioning
              # did = data['did']
              rev = data['commit']['rev']
                  
              # Serialize the event
              serialized = topic.serialize(key=rev, value=data)
              # Produce to Kafka
              with app.get_producer() as producer:
                  producer.produce(
                      topic=topic.name,
                      key=serialized.key,   
                      value=serialized.value
                  )

              # print(serialized)

            except websockets.ConnectionClosed as e:
              print(f"Connection closed: {e}")
              break
            except Exception as e:
              print(f"Error processing message: {e}")
              continue

      except Exception as e:
        print(f"Connection error: {e}. Reconnecting in 5 seconds...")
        await asyncio.sleep(5)


if __name__ == "__main__":
    try:
        asyncio.run(listen_to_bluesky())
    except KeyboardInterrupt:
        print("\n\nStopping stream...")

