requirements.txt
asyncio
quixstreams
websocket-client
